+++ tests/test_options.py
@@ -1,10 +1,13 @@
+from pathlib import Path
+from textwrap import dedent
+
 import pytest
 
 from pytest_pyodide.hook import _filter_runtimes
 
 
 def test_dist_dir(pytester):
-    dist_dir = "dist"
+    dist_dir = str(Path("dist").resolve())
 
     pytester.makepyfile(
         f"""
@@ -14,7 +17,7 @@ def test_dist_dir(pytester):
         """
     )
 
-    result = pytester.runpytest("--dist-dir", dist_dir)
+    result = pytester.runpytest("--dist-dir", "dist")
     result.assert_outcomes(passed=1)
 
 
@@ -86,3 +89,36 @@ def test_invalid_runtime(pytester, _runtime):
 )
 def test_filter_runtimes(_runtime, expected):
     assert _filter_runtimes(_runtime) == expected
+
+
+def test_options_pytester(pytester):
+    pytester.makepyfile(
+        dedent(
+            """
+            import pytest
+            from pathlib import Path
+
+            def test_options_pytester():
+                assert pytest.pyodide_run_host_test == True
+                assert pytest.pyodide_runtimes == set(["chrome","firefox","safari","node"])
+                assert pytest.pyodide_dist_dir == Path("some_weird_dir").resolve()
+            """
+        )
+    )
+    run_host = pytest.pyodide_run_host_test
+    runtimes = pytest.pyodide_runtimes
+    dist_dir = pytest.pyodide_dist_dir
+
+    result = pytester.runpytest(
+        "--dist-dir",
+        Path(__file__).parents[1] / "pyodide",
+        "--rt",
+        "chrome,firefox,safari,node",
+        "--dist-dir",
+        "some_weird_dir",
+    )
+    result.assert_outcomes(passed=1)
+
+    assert run_host == pytest.pyodide_run_host_test
+    assert runtimes == pytest.pyodide_runtimes
+    assert dist_dir == pytest.pyodide_dist_dir